+++ Source/Shaders/CentralBodyVS.glsl
@@ -1,4 +1,4 @@
-attribute vec3 position3D;
+attribute vec4 position3DAndHeight;
 attribute vec2 textureCoordinates;
 
 uniform float u_morphTime;
@@ -22,7 +22,7 @@ float get2DYPositionFraction();
 
 vec4 getPosition3DMode(vec3 position3DWC)
 {
-    return czm_projection * (u_modifiedModelView * vec4(position3D, 1.0));
+    return czm_projection * (u_modifiedModelView * vec4(position3DAndHeight.xyz, 1.0));
 }
 
 float get2DMercatorYPositionFraction()
@@ -54,16 +54,21 @@ float get2DGeographicYPositionFraction()
     return textureCoordinates.y;
 }
 
-vec4 getPosition2DMode(vec3 position3DWC)
+vec4 getPositionPlanarEarth(vec3 position3DWC, float height2D)
 {
     float yPositionFraction = get2DYPositionFraction();
-    vec4 rtcPosition2D = vec4(0.0, mix(u_tileExtent.st, u_tileExtent.pq, vec2(textureCoordinates.x, yPositionFraction)), 1.0);  
+    vec4 rtcPosition2D = vec4(height2D, mix(u_tileExtent.st, u_tileExtent.pq, vec2(textureCoordinates.x, yPositionFraction)), 1.0);  
     return czm_projection * (u_modifiedModelView * rtcPosition2D);
+}
+
+vec4 getPosition2DMode(vec3 position3DWC)
+{
+    return getPositionPlanarEarth(position3DWC, 0.0);
 }
 
 vec4 getPositionColumbusViewMode(vec3 position3DWC)
 {
-    return getPosition2DMode(position3DWC);
+    return getPositionPlanarEarth(position3DWC, position3DAndHeight.w);
 }
 
 vec4 getPositionMorphingMode(vec3 position3DWC)
@@ -78,11 +83,11 @@ vec4 getPositionMorphingMode(vec3 position3DWC)
 
 void main() 
 {
-    vec3 position3DWC = position3D + u_center3D;
+    vec3 position3DWC = position3DAndHeight.xyz + u_center3D;
 
     gl_Position = getPosition(position3DWC);
     
-    v_positionEC = (czm_modelView * vec4(position3DWC, 1.0)).xyz;
+    v_positionEC = (czm_modelView3D * vec4(position3DWC, 1.0)).xyz;
     v_positionMC = position3DWC;                                 // position in model coordinates
     v_textureCoordinates = textureCoordinates;
 }