+++ Specs/Scene/EllipsoidTerrainProviderSpec.js
@@ -28,41 +28,15 @@ defineSuite([
         destroyContext(context);
     });
 
-    it('creates a vertex array and bounding sphere for the tile.', function() {
-        var terrain;
-        var tile;
+    it('requestTileGeometry creates terrain data.', function() {
+        var terrain = new EllipsoidTerrainProvider();
+        var terrainData = terrain.requestTileGeometry(0, 0, 0);
+        expect(terrainData).toBeDefined();
+    });
 
-        runs(function() {
-            terrain = new EllipsoidTerrainProvider();
-            var tiles = terrain.tilingScheme.createLevelZeroTiles();
-            tile = tiles[0];
-
-            terrain.requestTileGeometry(tile);
-        });
-
-        waitsFor(function() {
-            return tile.state === TileState.RECEIVED;
-        });
-
-        runs(function() {
-            terrain.transformGeometry(context, tile);
-        });
-
-        waitsFor(function() {
-            return tile.state === TileState.TRANSFORMED;
-        });
-
-        runs(function() {
-            terrain.createResources(context, tile);
-        });
-
-        waitsFor(function() {
-            return tile.state === TileState.READY;
-        });
-
-        runs(function() {
-            expect(tile.vertexArray).not.toBeNull();
-            expect(tile.boundingSphere3D).not.toBeNull();
-        });
+    it('has error event', function() {
+        var provider = new EllipsoidTerrainProvider();
+        expect(provider.getErrorEvent()).toBeDefined();
+        expect(provider.getErrorEvent()).toBe(provider.getErrorEvent());
     });
 }, 'WebGL');